Small final updates for the routing extension, including:     - Minor docs update     - Rename extension function to be routing-extension-specific
