+++ vic/extensions/rout_rvic/include/rout.h
@@ -82,12 +82,12 @@ void gather_var_double(double *, double *);
  *****************************************************************************/
 void get_global_param_rout(FILE *gp);
 void cshift(double *, int, int, int, int);
-void vic_store_extension(nc_file_struct *);
-void vic_restore_extension(nameid_struct *, metadata_struct *);
-void state_metadata_extension();
-void set_nc_state_file_info_extension(nc_file_struct *);
-void set_nc_state_var_info_extension(nc_file_struct *);
-void initialize_state_file_extension(char *, nc_file_struct *);
+void vic_store_rout_extension(nc_file_struct *);
+void vic_restore_rout_extension(nameid_struct *, metadata_struct *);
+void state_metadata_rout_extension();
+void set_nc_state_file_info_rout_extension(nc_file_struct *);
+void set_nc_state_var_info_rout_extension(nc_file_struct *);
+void initialize_state_file_rout_extension(char *, nc_file_struct *);
 
 /******************************************************************************
  * @brief   Output state variable.
