+++ vic/extensions/rout_rvic/src/vic_restore_extension.c
@@ -31,8 +31,8 @@
  * @brief    Save model state.
  *****************************************************************************/
 void
-vic_restore_extension(nameid_struct   *init_state_file,
-                      metadata_struct *state_metadata)
+vic_restore_rout_extension(nameid_struct   *init_state_file,
+                           metadata_struct *state_metadata)
 {
     extern int         mpi_rank;
     extern rout_struct rout;
