+++ vic/extensions/rout_rvic/src/vic_store_extension.c
@@ -31,7 +31,7 @@
  * @brief    Save model state.
  *****************************************************************************/
 void
-vic_store_extension(nc_file_struct *nc_state_file)
+vic_store_rout_extension(nc_file_struct *nc_state_file)
 {
     extern int         mpi_rank;
     extern rout_struct rout;
@@ -60,7 +60,7 @@ vic_store_extension(nc_file_struct *nc_state_file)
  * @brief   Setup state file netcdf structure
  *****************************************************************************/
 void
-set_nc_state_file_info_extension(nc_file_struct *nc_state_file)
+set_nc_state_file_info_rout_extension(nc_file_struct *nc_state_file)
 {
     extern rout_struct rout;
 
@@ -77,7 +77,7 @@ set_nc_state_file_info_extension(nc_file_struct *nc_state_file)
  * @brief   Setup state variable dimensions, types, etc.
  *****************************************************************************/
 void
-set_nc_state_var_info_extension(nc_file_struct *nc)
+set_nc_state_var_info_rout_extension(nc_file_struct *nc)
 {
     size_t i;
     size_t j;
@@ -116,8 +116,8 @@ set_nc_state_var_info_extension(nc_file_struct *nc)
             and adding metadata.
  *****************************************************************************/
 void
-initialize_state_file_extension(char           *filename,
-                                nc_file_struct *nc_state_file)
+initialize_state_file_rout_extension(char           *filename,
+                                     nc_file_struct *nc_state_file)
 {
     int status;
