+++ vic/extensions/rout_stub/include/rout.h
@@ -55,12 +55,12 @@ void rout_alloc(void);      // allocate memory
 void rout_init(void);       // initialize model parameters from parameter files
 void rout_run(void);        // run routing over the domain
 void rout_finalize(void);   // clean up routine for routing
-void vic_store_extension(nc_file_struct *);
-void vic_restore_extension(nameid_struct *, metadata_struct *);
-void state_metadata_extension();
-void set_nc_state_file_info_extension(nc_file_struct *);
-void set_nc_state_var_info_extension(nc_file_struct *);
-void initialize_state_file_extension(char *, nc_file_struct *);
+void vic_store_rout_extension(nc_file_struct *);
+void vic_restore_rout_extension(nameid_struct *, metadata_struct *);
+void state_metadata_rout_extension();
+void set_nc_state_file_info_rout_extension(nc_file_struct *);
+void set_nc_state_var_info_rout_extension(nc_file_struct *);
+void initialize_state_file_rout_extension(char *, nc_file_struct *);
 
 /******************************************************************************
  * @brief   Output state variable.
