+++ vic/extensions/rout_stub/src/vic_restore_extension.c
@@ -31,7 +31,7 @@
  * @brief    Save model state.
  *****************************************************************************/
 void
-vic_restore_extension(nameid_struct   *init_state_file,
-                      metadata_struct *state_metadata)
+vic_restore_rout_extension(nameid_struct   *init_state_file,
+                           metadata_struct *state_metadata)
 {
 }
