+++ vic/extensions/rout_stub/src/vic_store_extension.c
@@ -31,7 +31,7 @@
  * @brief    Save model state.
  *****************************************************************************/
 void
-vic_store_extension(nc_file_struct *nc_state_file)
+vic_store_rout_extension(nc_file_struct *nc_state_file)
 {
 }
 
@@ -39,7 +39,7 @@ vic_store_extension(nc_file_struct *nc_state_file)
  * @brief   Setup state file netcdf structure
  *****************************************************************************/
 void
-set_nc_state_file_info_extension(nc_file_struct *nc_state_file)
+set_nc_state_file_info_rout_extension(nc_file_struct *nc_state_file)
 {
 }
 
@@ -47,7 +47,7 @@ set_nc_state_file_info_extension(nc_file_struct *nc_state_file)
  * @brief   Setup state variable dimensions, types, etc.
  *****************************************************************************/
 void
-set_nc_state_var_info_extension(nc_file_struct *nc)
+set_nc_state_var_info_rout_extension(nc_file_struct *nc)
 {
 }
 
@@ -56,7 +56,7 @@ set_nc_state_var_info_extension(nc_file_struct *nc)
             and adding metadata.
  *****************************************************************************/
 void
-initialize_state_file_extension(char           *filename,
-                                nc_file_struct *nc_state_file)
+initialize_state_file_rout_extension(char           *filename,
+                                     nc_file_struct *nc_state_file)
 {
 }
